Finish style tweaking for iframe_area
This mainly involved turning off the default border around the iframe
and suppressing any scrollbars.

The latter is tricky. The iframe scrolling=no attribute was removed
in HTML 5. The suggested replacement (the CSS 'overflow: hidden' CSS
attribute) only works when applied to the source page, not the enclosing
iframe ad typically you don't have control of the source page. This
patch pragmatically applies scrolling=no to the iframe.