Rework date handling to use UK localtime where needed
Times (as opposed to DateTimes) sent to and received from the
timetable API have to be in UK local time which may or may not be
the same a browser local time. likewise date-based times used
as filters for the real-time interface need to be ISO 8859-1 formatted
UK local times (with offset showun in seconds, not as Z).

This edit should ensure that UK local times are used where needed.
Time (in the form of the next move to British Summer Time) will
tell...