Proper fix for problem with with undefined delay
Edit a00e5c06 refered to entry.delay in display_nextbus()
when it should have been journey.delay (cut-and-paste error from
elsewhere where the variable is called 'entry'). This resulted in

  Cannot read property 'asMinutes' of undefined

becasue, while there *was* an entry variable in scope, it was the
wrong sort of thing and didn't have a delay property.

This should provide a perminant fix.