Add support for showing several traffic maps in sequence
Traffic maps take up a lot of space, but it's often useful to show
several. This edit allows the TrafficMap widgit to show several
maps in a rotating sequence.

The widget is now configured with an array of map definitions - at
least one is required. There's some support for backward
compatibility that accepts a single map definition as the entire
set of parameters.

The rotation uses Google facilities to re-centre and/or re-zoom
the map and in many cases Google chooses to animate the transition
which helps to preserve context.

REvides version using Google's API to swap maps