Fix daily reload; improve logging; further shrink 'Park & Ride'
Switch the daily reload from using setTimeout() to using setInterval() -
it appears that a setTimeout() event can be lost of the browser is
asleep when it is supposed to be run. The new code works out when the
next reload should happen, and then checks every minute to see if that
time has passed. This should be safe even in the face of the browser
being asleep some of the time.

Modify log() to prefix messages with the name of the container. Do
the same by hand for a couple of console.log() calls.

Abbreviate all the P&R lines to just 'P&R' without the number to save
a bit more screen space. With luck, it will be enough for passengers
to know it's *a* P&R bus but not which one.