Further weather fix to stop 'Now (15:00)'' appearing before 'Lunchtime'
At 13:45 the 'now' forecast is the one for 15:00 but the 'Lunchtime'
one for 12:00 is still being displayed. Previously, the 'current'
forecast was only suppressed if it was the same as the topmost other
one. This patch suppresses the 'current' forecast if it doesn't
have a timestamp that is earlier than the topmost other one.